+++ addons/web/views/webclient_templates.xml
@@ -625,6 +625,9 @@
                 <script type="text/javascript" src="/web/static/tests/chrome/search_view_mobile_tests.js"></script>
                 <script type="text/javascript" src="/web/static/tests/views/calendar_mobile_tests.js"></script>
                 <script type="text/javascript" src="/web/static/tests/views/kanban_mobile_tests.js"></script>
+                <script type="text/javascript" src="/web/static/tests/views/form_mobile_tests.js"></script>
+                <script type="text/javascript" src="/web/static/tests/views/basic_fields_mobile_tests.js"></script>
+                <script type="text/javascript" src="/web/static/tests/views/relational_fields_mobile_tests.js"></script>
             </t>
 
             <div id="qunit"/>
